+++ packages/aws-durable-execution-sdk-js/src/types/index.ts
@@ -19,16 +19,6 @@ export enum DurableExecutionMode {
   ReplaySucceededContext = "ReplaySucceededContext",
 }
 
-// Import types for concurrent execution
-import type {
-  ConcurrentExecutionItem,
-  ConcurrentExecutor,
-  ConcurrencyConfig,
-} from "../handlers/concurrent-execution-handler/concurrent-execution-handler";
-
-// Import BatchResult types
-import type { BatchResult } from "../handlers/concurrent-execution-handler/batch-result";
-
 export interface LambdaHandler<T> {
   (event: T, context: Context): Promise<DurableExecutionInvocationOutput>;
 }
@@ -427,7 +417,7 @@ export interface DurableContext {
     name: string | undefined,
     items: TInput[],
     mapFunc: MapFunc<TInput, TOutput>,
-    config?: MapConfig<TInput>,
+    config?: MapConfig<TInput, TOutput>,
   ): Promise<BatchResult<TOutput>>;
 
   /**
@@ -446,7 +436,7 @@ export interface DurableContext {
   map<TInput, TOutput>(
     items: TInput[],
     mapFunc: MapFunc<TInput, TOutput>,
-    config?: MapConfig<TInput>,
+    config?: MapConfig<TInput, TOutput>,
   ): Promise<BatchResult<TOutput>>;
   /**
    * Executes multiple functions in parallel with optional concurrency control
@@ -468,7 +458,7 @@ export interface DurableContext {
   parallel<T>(
     name: string | undefined,
     branches: (ParallelFunc<T> | NamedParallelBranch<T>)[],
-    config?: ParallelConfig,
+    config?: ParallelConfig<T>,
   ): Promise<BatchResult<T>>;
 
   /**
@@ -487,7 +477,7 @@ export interface DurableContext {
    */
   parallel<T>(
     branches: (ParallelFunc<T> | NamedParallelBranch<T>)[],
-    config?: ParallelConfig,
+    config?: ParallelConfig<T>,
   ): Promise<BatchResult<T>>;
   promise: {
     /**
@@ -740,7 +730,7 @@ export interface DurableContext {
     name: string | undefined,
     items: ConcurrentExecutionItem<TItem>[],
     executor: ConcurrentExecutor<TItem, TResult>,
-    config?: ConcurrencyConfig,
+    config?: ConcurrencyConfig<TResult>,
   ): Promise<BatchResult<TResult>>;
 
   /**
@@ -761,7 +751,7 @@ export interface DurableContext {
   executeConcurrently<TItem, TResult>(
     items: ConcurrentExecutionItem<TItem>[],
     executor: ConcurrentExecutor<TItem, TResult>,
-    config?: ConcurrencyConfig,
+    config?: ConcurrencyConfig<TResult>,
   ): Promise<BatchResult<TResult>>;
 
   /**
@@ -1002,11 +992,15 @@ export interface WaitForConditionConfig<T> {
 /**
  * Configuration options for map operations
  */
-export interface MapConfig<T> {
+export interface MapConfig<TItem, TResult> {
   /** Maximum number of concurrent executions (default: unlimited) */
   maxConcurrency?: number;
   /** Function to generate custom names for map items */
-  itemNamer?: (item: T, index: number) => string;
+  itemNamer?: (item: TItem, index: number) => string;
+  /** Serialization/deserialization configuration for parent context */
+  serdes?: Serdes<BatchResult<TResult>>;
+  /** Serialization/deserialization configuration for each item */
+  itemSerdes?: Serdes<TResult>;
   /** Configuration for completion behavior */
   completionConfig?: {
     /** Minimum number of successful executions required */
@@ -1021,9 +1015,13 @@ export interface MapConfig<T> {
 /**
  * Configuration options for parallel operations
  */
-export interface ParallelConfig {
+export interface ParallelConfig<TResult> {
   /** Maximum number of concurrent executions (default: unlimited) */
   maxConcurrency?: number;
+  /** Serialization/deserialization configuration for parent context */
+  serdes?: Serdes<BatchResult<TResult>>;
+  /** Serialization/deserialization configuration for each branch */
+  itemSerdes?: Serdes<TResult>;
   /** Configuration for completion behavior */
   completionConfig?: {
     /** Minimum number of successful executions required */
@@ -1035,15 +1033,106 @@ export interface ParallelConfig {
   };
 }
 
-// Re-export concurrent execution types for public API
-export type {
-  ConcurrentExecutionItem,
-  ConcurrentExecutor,
-  ConcurrencyConfig,
-} from "../handlers/concurrent-execution-handler/concurrent-execution-handler";
+/**
+ * Represents an item to be executed with metadata for deterministic replay
+ */
+export interface ConcurrentExecutionItem<T> {
+  /** Unique identifier for the item */
+  id: string;
+  /** The actual data/payload for the item */
+  data: T;
+  /** Index of the item in the original array */
+  index: number;
+  /** Optional custom name for the item */
+  name?: string;
+}
 
-// Re-export BatchResult types for public API
-export type {
-  BatchResult,
-  BatchItem,
-} from "../handlers/concurrent-execution-handler/batch-result";
+/**
+ * Executor function type for concurrent execution
+ */
+export type ConcurrentExecutor<TItem, TResult> = (
+  item: ConcurrentExecutionItem<TItem>,
+  childContext: DurableContext,
+) => Promise<TResult>;
+
+/**
+ * Represents a single item in a batch result
+ */
+export interface BatchItem<R> {
+  /** The result value if the item succeeded */
+  result?: R;
+  /** The error if the item failed */
+  error?: Error;
+  /** Index of the item in the original array */
+  index: number;
+  /** Status of the item execution */
+  status: BatchItemStatus;
+}
+
+/**
+ * Result of a batch operation (map, parallel, or concurrent execution)
+ */
+export interface BatchResult<R> {
+  /** All items in the batch with their results/errors */
+  all: Array<BatchItem<R>>;
+
+  /** Returns only the items that succeeded */
+  succeeded(): Array<BatchItem<R> & { result: R }>;
+  /** Returns only the items that failed */
+  failed(): Array<BatchItem<R> & { error: Error }>;
+  /** Returns only the items that are still in progress */
+  started(): Array<BatchItem<R> & { status: BatchItemStatus.STARTED }>;
+
+  /** Overall status of the batch (SUCCEEDED if no failures, FAILED otherwise) */
+  status: BatchItemStatus.SUCCEEDED | BatchItemStatus.FAILED;
+  /** Reason why the batch completed */
+  completionReason:
+    | "ALL_COMPLETED"
+    | "MIN_SUCCESSFUL_REACHED"
+    | "FAILURE_TOLERANCE_EXCEEDED";
+  /** Whether any item in the batch failed */
+  hasFailure: boolean;
+
+  /** Throws the first error if any item failed */
+  throwIfError(): void;
+  /** Returns array of all successful results */
+  getResults(): Array<R>;
+  /** Returns array of all errors */
+  getErrors(): Array<Error>;
+
+  /** Number of successful items */
+  successCount: number;
+  /** Number of failed items */
+  failureCount: number;
+  /** Number of started but not completed items */
+  startedCount: number;
+  /** Total number of items */
+  totalCount: number;
+}
+
+/**
+ * Configuration options for concurrent execution operations
+ */
+export interface ConcurrencyConfig<TResult> {
+  /** Maximum number of concurrent executions (default: unlimited) */
+  maxConcurrency?: number;
+  /** Top-level operation subtype for tracking */
+  topLevelSubType?: string;
+  /** Iteration-level operation subtype for tracking */
+  iterationSubType?: string;
+  /** Function to generate summary from batch result */
+  summaryGenerator?: (result: BatchResult<TResult>) => string;
+  /** Serialization/deserialization configuration for parent context */
+  serdes?: Serdes<BatchResult<TResult>>;
+  /** Serialization/deserialization configuration for each item */
+  itemSerdes?: Serdes<TResult>;
+  /** Configuration for completion behavior */
+  completionConfig?: {
+    /** Minimum number of successful executions required */
+    minSuccessful?: number;
+    /** Maximum number of failures tolerated */
+    toleratedFailureCount?: number;
+    /** Maximum percentage of failures tolerated (0-100) */
+    toleratedFailurePercentage?: number;
+  };
+}
